requirements.txt
csaps==1.1.0
matplotlib~=3.8.2
numpy~=1.26.3
pandas~=2.2.0
customtkinter~=5.2.2
pyserial
matplotlib
openpyxl

